+++ tests/Unit/PrependTest.php
@@ -6,41 +6,41 @@ use Okipa\LaravelTable\Table;
 use Okipa\LaravelTable\Test\Models\User;
 use Okipa\LaravelTable\Test\LaravelTableTestCase;
 
-class IconTest extends LaravelTableTestCase
+class PrependTest extends LaravelTableTestCase
 {
-    public function testSetIconAttribute()
+    public function testSetPrependAttribute()
     {
         $table = (new Table)->model(User::class);
-        $table->column('name')->icon('icon');
-        $this->assertEquals('icon', $table->columns->first()->icon);
-        $this->assertEquals(false, $table->columns->first()->displayIconWhenNoValue);
+        $table->column('name')->prepend('icon');
+        $this->assertEquals('icon', $table->columns->first()->prepend);
+        $this->assertEquals(false, $table->columns->first()->displayPrependEvenIfNoValue);
     }
 
-    public function testSetIconAttributeAndSetShowWithNoValue()
+    public function testSetPrependAttributeAndSetShowWithNoValue()
     {
         $table = (new Table)->model(User::class);
-        $table->column('name')->icon('icon', true);
-        $this->assertEquals('icon', $table->columns->first()->icon);
-        $this->assertEquals(true, $table->columns->first()->displayIconWhenNoValue);
+        $table->column('name')->prepend('icon', true);
+        $this->assertEquals('icon', $table->columns->first()->prepend);
+        $this->assertEquals(true, $table->columns->first()->displayPrependEvenIfNoValue);
     }
 
-    public function testSetIconHtml()
+    public function testSetPrependHtml()
     {
         $this->createMultipleUsers(1);
         $this->routes(['users'], ['index']);
         $table = (new Table)->model(User::class)->routes(['index' => ['name' => 'users.index']]);
-        $table->column('name')->icon('icon');
+        $table->column('name')->prepend('icon');
         $table->render();
         $html = view('laravel-table::' . $table->tbodyTemplatePath, compact('table'))->render();
         $this->assertStringContainsString('icon', $html);
     }
 
-    public function testSetIconWithCustomValueHtml()
+    public function testSetPrependWithCustomValueHtml()
     {
         $this->createMultipleUsers(1);
         $this->routes(['users'], ['index']);
         $table = (new Table)->model(User::class)->routes(['index' => ['name' => 'users.index']]);
-        $table->column('name')->icon('icon')->value(function () {
+        $table->column('name')->prepend('icon')->value(function () {
             return 'test';
         });
         $table->render();
@@ -48,25 +48,25 @@ class IconTest extends LaravelTableTestCase
         $this->assertStringContainsString('icon', $html);
     }
 
-    public function testSetIconWithNoValueHtml()
+    public function testSetPrependWithNoValueHtml()
     {
         $user = $this->createUniqueUser();
         $user->update(['name' => null]);
         $this->routes(['users'], ['index']);
         $table = (new Table)->model(User::class)->routes(['index' => ['name' => 'users.index']]);
-        $table->column('name')->icon('icon');
+        $table->column('name')->prepend('icon');
         $table->render();
         $html = view('laravel-table::' . $table->tbodyTemplatePath, compact('table'))->render();
         $this->assertStringNotContainsString('icon', $html);
     }
 
-    public function testSetIconWithNoValueButShowAnywayValueHtml()
+    public function testSetPrependWithNoValueButShowAnywayValueHtml()
     {
         $user = $this->createUniqueUser();
         $user->update(['name' => null]);
         $this->routes(['users'], ['index']);
         $table = (new Table)->model(User::class)->routes(['index' => ['name' => 'users.index']]);
-        $table->column('name')->icon('icon', true);
+        $table->column('name')->prepend('icon', true);
         $table->render();
         $html = view('laravel-table::' . $table->tbodyTemplatePath, compact('table'))->render();
         $this->assertStringContainsString('icon', $html);